test(core): isolate timeout probe timings
Keep only the two process-launch survival probes in the nonparallel process collection while leaving atomic marker support tests parallelizable.

Copilot CLI: 1.0.80
Model: GPT-5.6 Sol (gpt-5.6-sol)
